move examples out of src
