+++ examples/basic.ts
@@ -2,9 +2,9 @@ import { WidgetBase } from '@dojo/widget-core/WidgetBase';
 import { v, w } from '@dojo/widget-core/d';
 import { WidgetProperties, DNode } from '@dojo/widget-core/interfaces';
 
-import { Outlet } from './../Outlet';
-import { Link } from './../Link';
-import { MapParamsOptions } from './../interfaces';
+import { Outlet } from './../src/Outlet';
+import { Link } from './../src/Link';
+import { MapParamsOptions } from './../src/interfaces';
 
 export interface ChildProperties extends WidgetProperties {
 	name: string;
