+++ examples/main.ts
@@ -3,9 +3,9 @@ import { v, w } from '@dojo/widget-core/d';
 import { ProjectorMixin } from '@dojo/widget-core/mixins/Projector';
 import { Registry } from '@dojo/widget-core/Registry';
 
-import { Config } from './../interfaces';
-import { registerRouterInjector } from './../RouterInjector';
-import { Link } from './../Link';
+import { Config } from './../src/interfaces';
+import { registerRouterInjector } from './../src/RouterInjector';
+import { Link } from './../src/Link';
 import { BasicAppOutlet, BasicAppRouteConfig } from './basic';
 import { UrlParametersAppOutlet, UrlParametersRouteConfig } from './url-parameters';
 import { AmbiguousMatchesOutlet, AmbiguousMatchesRouteConfig } from './ambigious-matches';
